fixed shop class and added list rather than map
