+++ src/main/java/models/Shop.java
@@ -1,0 +1,43 @@
+package models;
+
+import javax.persistence.*;
+import java.util.List;
+
+@Entity
+@Table(name="shop")
+public class Shop {
+
+
+    private int id;
+    private List<Product> stock;
+
+    public Shop(List<Product> stock) {
+        this.stock = stock;
+    }
+
+    public Shop() {
+    }
+
+
+
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "id")
+    public int getId() {
+        return id;
+    }
+
+    public void setId(int id) {
+        this.id = id;
+    }
+
+    @Column(name = "stock")
+    public List<Product> getStock() {
+        return stock;
+    }
+
+    public void setStock(List<Product> stock) {
+        this.stock = stock;
+    }
+}
